+++ CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet!
+- Add support for `prettier-plugin-multiline-arrays` ([#299](https://github.com/tailwindlabs/prettier-plugin-tailwindcss/pull/299))
 
 ## [0.6.5] - 2024-06-17
 
-- Only re-apply string escaping when necessary  ([#295](https://github.com/tailwindlabs/prettier-plugin-tailwindcss/pull/295))
+- Only re-apply string escaping when necessary ([#295](https://github.com/tailwindlabs/prettier-plugin-tailwindcss/pull/295))
 
 ## [0.6.4] - 2024-06-12